:umbrella: (eslint) separate linter from utils test suite
This is to avoid overloading and time wasting
